WIP: K8s (#49)
* Add broken test

* Update Ci pipeline

* Add CI caching to speed up the build

* Add test coverage check and unittest parser

* Add token to unittest result publisher

* Ensure test publisher uses correct path

* Add initial flyway migration

* Add docker image build and registry push to CI

* Add frontend build & deploy

* Add pathbased triggers for frontend build

* Fix image push secret

* Add manual trigger to frontend deploy

* U[date frontend deploy to use correct paths

* Add initial kubernetes automation

* Add flux v0.11.0 components manifests

* Add flux v0.11.0 sync manifests

* Add ssss and ambassador

* Add flux v0.11.0 components manifests

* Add flux v0.11.0 sync manifests

* Add flux v0.11.0 components manifests

* Add flux v0.11.0 sync manifests

* Add flux v0.11.0 components manifests

* Add flux v0.11.0 sync manifests

* Add flux v0.11.0 components manifests

* Add flux v0.11.0 sync manifests

* Add flux v0.11.0 components manifests

* Add flux v0.11.0 sync manifests

* Add flux v0.11.0 components manifests

* Add flux v0.11.0 sync manifests

* fix ssss deployment

* fix ssss deployment

* fix ssss deployment

* Add ambassador kustomization

* remove old files

* add kustomizations

* add kustomizations

* remove unused label

* remove unused label

* Add image pull secret

* use traefic ingress set up by k3s

* use traefic ingress set up by k3s

* increase java memory limit

* Add sealed secrets

* remove sealed secrets dependency

* add login/signup/oauth (#15)

* add login/signup/oauth

* Update build.gradle

* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/AppProperties.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/AppProperties.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/AppProperties.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/AppProperties.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/Secrets.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/WebSecurityConfig.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/dto/LocalUser.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/dto/LoginRequest.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/dto/SocialProvider.java


* Feedback verwerkt

* Added extra tests/improvements PP session with David

* Fix sonarlint issues

* Lombokify classes for less testing load

* Lombokify classes for less testing load

* Lombokify classes for less testing load

* fix cyclic dependency

* Fix h2 in memory database

* Change port to 8080

* Take coveraget to 100%

* Add secrets.properties to build

* Update jib and remove sonarlint


* #4 implemented and rebased (#21)


* Forgot password mail signup (#22)

* add login/signup/oauth

* Update build.gradle

* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/AppProperties.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/AppProperties.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/AppProperties.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/AppProperties.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/Secrets.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/configuration/WebSecurityConfig.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/dto/LocalUser.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/dto/LoginRequest.java


* Update backend/src/main/java/nl/jed/supersimplesupplysystem/dto/SocialProvider.java


* Feedback verwerkt

* Added extra tests/improvements PP session with David

* Fix sonarlint issues

* Lombokify classes for less testing load

* Lombokify classes for less testing load

* Lombokify classes for less testing load

* fix cyclic dependency

* Fix h2 in memory database

* Change port to 8080

* #4 - Added H2 console functionalities

* #4 - added classpath to .gitignore

* #4 - Household entity, controller, service, repository are built

* #4 - CRUD Operations now complete

* #4 - Added Mockito Dependencies, Tests are written

* #4 - Created frontend basic UI, added households page. Login and register left to do

* #4 - Login Screen Ui now designed and works.

* #4 - Registration Screen Ui now designed and works.

* #4 - Registration Screen Ui now designed and works.

* #4 - Registration Screen Ui now designed and works.

* Started implementing reset password

* Forgot password + Email verification flow implemented

* Wrote some tests

* Deleted controller tests


* Fixed forgot password flow (#23)

* Fix for certificate (#24)


* #18 (#26)

* #18 adding households in ui now implemented

* #18 - Adding Households Through UI now works

* #18 - Added ManyToMany Relations between users and households.

* #18 - WIP Household Pre or PostFilters

* #18 - Implementation

* #18 - Implementation front and backend

* #18 - Implementation of filtering households per user. Also added UI funcitonality to build households.


* #11 - Implementation PWA and EAN scanner (#27)


* Create codeql-analysis.yml (#30)

* Create codeql-analysis.yml

* Add mising key to build

* fix key path

* Connected product front and backend to use scanner and get info (#32)


* remove coverage limit from ci step (#37)

* Main reset to all latest versions (#41)

Latest version

* removed outer files (#42)


* Added Swagger Config and whitelisted. (#43)


* Add products #14 #34 (#44)

* Gestart met producten toevoegen

* Implemented CRUD products

* Add broken test

* Add test coverage check and unittest parser

* Add token to unittest result publisher

* Add docker image build and registry push to CI

* Add frontend build & deploy

* remove sealed secrets

* Add sealed secrets

* add convenience script for refreshing secrets

* update secrets.yaml

* update secrets.yaml

* Update secrets

* move secrets to work with k8s secret mounting

* add certmanager and default base path

* cert-manager to sync

* Fix TempOntwikkelTest

* update jib to v3.0

* Update backend container version

* Add cluster issuer

* Add cluster issuer

* Add cluster issuer

* Add cluster issuer

* force ssl

* add correct selectors

* add correct selectors

* add correct path

+++ backend/src/main/java/nl/jed/supersimplesupplysystem/dto/openfoodfacts/Product.java
@@ -422,7 +422,7 @@ public class Product{
 	private String tracesFromUser;
 
 	@SerializedName("nutrition_score_warning_fruits_vegetables_nuts_estimate_from_ingredients_value")
-	private Integer nutritionScoreWarningFruitsVegetablesNutsEstimateFromIngredientsValue;
+	private Double nutritionScoreWarningFruitsVegetablesNutsEstimateFromIngredientsValue;
 
 	@SerializedName("countries_tags")
 	private List<String> countriesTags;
